# Remove pre-parse for array-based JSON