@iSazonov 2021-06-29
Have we direct test for the scenario?