@strawgate 2021-06-29
> Have we direct test for the scenario?

I added a test for this scenario