@daxian-dbw 2022-07-25
The pre-parsing was previously removed by https://github.com/PowerShell/PowerShell/pull/8346, but then a regression was found because a feature-level test failed: https://github.com/PowerShell/PowerShell/pull/8346#issuecomment-443326406, and hence the change was reverted (https://github.com/PowerShell/PowerShell/pull/8375).

I have verified that the previously failing feature-level test passes with changes in this PR. So, it looks to me we are safe to merge this PR.
https://github.com/PowerShell/PowerShell/blob/3d4e294262d2f2d7472ebd5d849b1b68fd6ef524/test/powershell/Modules/Microsoft.PowerShell.Utility/Json.Tests.ps1#L1698-L1711

![image](https://user-images.githubusercontent.com/127450/180870566-f20f58b0-86fe-43a9-b406-0977b2622677.png)